Move state decorators to utils.decorators.state
